requirements.txt
streamlit
google-generativeai
PyPDF2
numpy
Pillow